+++ include/tty.h
@@ -9,9 +9,9 @@ typedef struct {
   struct termios og_tty;
 } tty_t;
 
-void                               tty_enable_raw_mode(void);
-void                               tty_disable_raw_mode(void);
-__attribute__((weak)) unsigned int tty_get_window_size(unsigned int* rows, unsigned int* cols);
-void                               tty_clear(void);
+void                         tty_enable_raw_mode(void);
+void                         tty_disable_raw_mode(void);
+__attribute__((weak)) size_t tty_get_window_size(size_t* rows, size_t* cols);
+void                         tty_clear(void);
 
 #endif /* TTY_H */
