+++ t/unit/file.test.c
@@ -3,8 +3,8 @@
 #include "keypress.h"
 #include "tests.h"
 
-unsigned int
-tty_get_window_size (unsigned int *rows, unsigned int *cols) {
+size_t
+tty_get_window_size (size_t *rows, size_t *cols) {
   *rows = 40;
   *cols = 50;
   return 0;
@@ -90,7 +90,7 @@ run_file_mgmt_tests (void) {
     test_editor_open,
   };
 
-  for (unsigned int i = 0; i < sizeof(functions) / sizeof(functions[0]); i++) {
+  for (size_t i = 0; i < sizeof(functions) / sizeof(functions[0]); i++) {
     setup();
     functions[i]();
     teardown();
